requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
google-generativeai>=0.1.0
PyMuPDF>=1.20.0
python-dotenv>=0.20.0
pandas>=1.4.0
tabulate>=0.8.9